+++ android/app/src/main/java/com/now/naaga/data/remote/dto/third/AdventureStatusDto.kt
@@ -1,4 +1,4 @@
-package com.now.naaga.data.remote.dto.second
+package com.now.naaga.data.remote.dto.third
 
 data class AdventureStatusDto(
     val gameStatus: String,
