+++ android/app/src/main/java/com/now/naaga/data/remote/dto/third/FinishGameDto.kt
@@ -1,11 +1,10 @@
-package com.now.naaga.data.remote.dto.second
+package com.now.naaga.data.remote.dto.third
 
 import com.now.naaga.data.remote.dto.CoordinateDto
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class HintDto(
-    val id: Long,
+data class FinishGameDto(
+    val endType: String,
     val coordinate: CoordinateDto,
-    val direction: String,
 )
