+++ android/app/src/main/java/com/now/naaga/data/remote/dto/third/PlayerDto.kt
@@ -1,4 +1,4 @@
-package com.now.naaga.data.remote.dto.second
+package com.now.naaga.data.remote.dto.third
 
 import kotlinx.serialization.Serializable
